Port upstream 0.54.0: decompose thermo hot spots
